fix: Move `api_server` into module under `firecracker`
