+++ src/firecracker/Cargo.toml
@@ -11,17 +11,21 @@ license = "Apache-2.0"
 [dependencies]
 event-manager = "0.3.0"
 libc = "0.2.117"
+serde = { version = "1.0.136", features = ["derive"] }
+serde_derive = "1.0.136"
 serde_json = "1.0.78"
 timerfd = "1.3.0"
+thiserror = "1.0.32"
 
-api_server = { path = "../api_server" }
 logger = { path = "../logger" }
 mmds = { path = "../mmds" }
 seccompiler = { path = "../seccompiler" }
 snapshot = { path = "../snapshot"}
 utils = { path = "../utils" }
 vmm = { path = "../vmm" }
+micro_http = { git = "https://github.com/firecracker-microvm/micro-http", rev = "4b18a04" }
 
 [dev-dependencies]
 cargo_toml = "0.13.0"
 regex = "1.5.5"
+libc = "0.2.117"
