+++ src/firecracker/src/api_server/request/actions.rs
@@ -5,10 +5,10 @@ use logger::{IncMetric, METRICS};
 use serde::{Deserialize, Serialize};
 
 use super::super::VmmAction;
-use crate::parsed_request::{Error, ParsedRequest};
-use crate::request::Body;
+use crate::api_server::parsed_request::{Error, ParsedRequest};
+use crate::api_server::request::Body;
 #[cfg(target_arch = "aarch64")]
-use crate::request::StatusCode;
+use crate::api_server::request::StatusCode;
 
 // The names of the members from this enum must precisely correspond (as a string) to the possible
 // values of "action_type" from the json request body. This is useful to get a strongly typed
