+++ src/firecracker/src/api_server/request/balloon.rs
@@ -7,8 +7,8 @@ use vmm::vmm_config::balloon::{
 };
 
 use super::super::VmmAction;
-use crate::parsed_request::{Error, ParsedRequest};
-use crate::request::Body;
+use crate::api_server::parsed_request::{Error, ParsedRequest};
+use crate::api_server::request::Body;
 
 pub(crate) fn parse_get_balloon(path_second_token: Option<&&str>) -> Result<ParsedRequest, Error> {
     match path_second_token {
@@ -52,7 +52,7 @@ pub(crate) fn parse_patch_balloon(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::parsed_request::tests::vmm_action_from_request;
+    use crate::api_server::parsed_request::tests::vmm_action_from_request;
 
     #[test]
     fn test_parse_get_balloon_request() {
