+++ src/firecracker/src/api_server/request/version.rs
@@ -4,7 +4,7 @@
 use logger::{IncMetric, METRICS};
 use vmm::rpc_interface::VmmAction;
 
-use crate::parsed_request::{Error, ParsedRequest};
+use crate::api_server::parsed_request::{Error, ParsedRequest};
 
 pub(crate) fn parse_get_version() -> Result<ParsedRequest, Error> {
     METRICS.get_api_requests.vmm_version_count.inc();
@@ -14,7 +14,7 @@ pub(crate) fn parse_get_version() -> Result<ParsedRequest, Error> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::RequestAction;
+    use crate::api_server::RequestAction;
 
     #[test]
     fn test_parse_get_version_request() {
