+++ src/firecracker/src/api_server/request/vsock.rs
@@ -5,8 +5,8 @@ use logger::{IncMetric, METRICS};
 use vmm::vmm_config::vsock::VsockDeviceConfig;
 
 use super::super::VmmAction;
-use crate::parsed_request::{Error, ParsedRequest};
-use crate::request::Body;
+use crate::api_server::parsed_request::{Error, ParsedRequest};
+use crate::api_server::request::Body;
 
 pub(crate) fn parse_put_vsock(body: &Body) -> Result<ParsedRequest, Error> {
     METRICS.put_api_requests.vsock_count.inc();
@@ -36,7 +36,7 @@ pub(crate) fn parse_put_vsock(body: &Body) -> Result<ParsedRequest, Error> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::parsed_request::tests::depr_action_from_req;
+    use crate::api_server::parsed_request::tests::depr_action_from_req;
 
     #[test]
     fn test_parse_put_vsock_request() {
