+++ src/firecracker/src/api_server/swagger/firecracker.yaml
